backend/requirements.txt

fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
gitpython==3.1.32
python-multipart==0.0.6
httpx==0.24.1
openai==0.28.0
tiktoken==0.5.1
pytest==7.4.0
pytest-asyncio==0.21.1